+++ src/main/java/com/kh/samdado/business/model/dao/businessDao.java
@@ -2,8 +2,6 @@ package com.kh.samdado.business.model.dao;
 
 import java.util.List;
 
-import com.kh.samdado.admin.model.vo.Report;
-import com.kh.samdado.business.model.vo.Alliance;
 import com.kh.samdado.business.model.vo.Jjim;
 import com.kh.samdado.business.model.vo.Review;
 import com.kh.samdado.business.model.vo.business.Business;
@@ -13,6 +11,8 @@ import com.kh.samdado.business.model.vo.rentcar.Car;
 import com.kh.samdado.business.model.vo.rentcar.CarAtt;
 import com.kh.samdado.business.model.vo.rentcar.CarBooking;
 import com.kh.samdado.business.model.vo.tour.TourBooking;
+import com.kh.samdado.common.model.vo.Alliance;
+import com.kh.samdado.common.model.vo.Report;
 
 public interface businessDao {
