+++ src/main/java/com/kh/samdado/business/model/service/businessService.java
@@ -2,8 +2,8 @@ package com.kh.samdado.business.model.service;
 
 import java.util.List;
 
-import com.kh.samdado.admin.model.vo.Report;
-import com.kh.samdado.business.model.vo.Alliance;
+import com.kh.samdado.common.model.vo.Report;
+import com.kh.samdado.common.model.vo.Alliance;
 import com.kh.samdado.business.model.vo.Jjim;
 import com.kh.samdado.business.model.vo.Review;
 import com.kh.samdado.business.model.vo.business.Business;
