+++ src/main/java/com/kh/samdado/mypage/model/dao/MypageDao.java
@@ -2,9 +2,9 @@ package com.kh.samdado.mypage.model.dao;
 
 import java.util.List;
 
-import com.kh.samdado.business.model.vo.Alliance;
 import com.kh.samdado.business.model.vo.Jjim;
 import com.kh.samdado.business.model.vo.business.Business;
+import com.kh.samdado.common.model.vo.Alliance;
 import com.kh.samdado.common.model.vo.Income;
 import com.kh.samdado.mypage.model.vo.AccountBook;
 import com.kh.samdado.mypage.model.vo.Alert;
